pyproject.toml
[project]
name = "datathon-sunass"
version = "0.1.0"
description = "Entrega Datathon SUNASS 2026 — Categoria I Operacional (UCSP, equipo SSA11)"
requires-python = ">=3.12,<3.13"
readme = "README.md"
authors = [{ name = "f3r21" }, { name = "fabio" }]

dependencies = [
    # IO y datos
    "polars>=1.12",
    "pandas>=2.2",
    "pyarrow>=17.0",
    "pyreadstat>=1.2",
    "openpyxl>=3.1",
    "xlsxwriter>=3.2",
    "python-dotenv>=1.0",
    "pyyaml>=6.0",
    # ML clasico
    "scikit-learn>=1.5",
    "xgboost>=2.1",
    "lightgbm>=4.5",
    "imbalanced-learn>=0.12",
    "statsmodels>=0.14",
    "shap>=0.46",
    "ruptures>=1.1",
    # Geo (mapa folium)
    "folium>=0.17",
    "shapely>=2.0",
    # Viz
    "matplotlib>=3.9",
    "plotly>=5.24",
    # Dashboard + auth
    "streamlit>=1.39",
    "streamlit-authenticator>=0.4",
    "bcrypt>=4.2",
    # Reportes (PDF + XLSX)
    "pymupdf>=1.24",
]

[project.optional-dependencies]
dev = ["ruff>=0.7", "mypy>=1.13"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[dependency-groups]
dev = [
    "jupyter-cache>=1.0.1",
    "ipykernel>=6.29",
]
